FNAV-1 - Filament 4 Upgrade, refactoring various items to be done in the new Filament 4 fashion
