+++ composer.json
@@ -25,7 +25,7 @@
         "php": "^8.1|^8.2",
         "doctrine/dbal": "^3.5.1",
         "filament/filament": "^3.0|^4.0",
-        "guava/filament-icon-picker": "^2.0",
+        "guava/filament-icon-picker": "^2.0|^3.0",
         "illuminate/contracts": "^10.12|^11.0|^12.0",
         "spatie/laravel-package-tools": "^1.9.2"
     },
