+++ resources/views/navigation-builder.blade.php
@@ -2,7 +2,6 @@
     :id="$getId()"
     :label="$getLabel()"
     :label-sr-only="$isLabelHidden()"
-    :helper-text="$getHelperText()"
     :hint="$getHint()"
     :hint-icon="$getHintIcon()"
     :required="$isRequired()"
