+++ src/Filament/Resources/NavigationResource/Pages/Concerns/HandlesNavigationBuilder.php
@@ -4,14 +4,15 @@ namespace Digitonic\FilamentNavigation\Filament\Resources\NavigationResource\Pag
 
 use Digitonic\FilamentNavigation\FilamentNavigation;
 use Filament\Actions\Action;
-use Filament\Forms\ComponentContainer;
-use Filament\Forms\Components\Component;
-use Filament\Forms\Components\Group;
 use Filament\Forms\Components\Select;
 use Filament\Forms\Components\TextInput;
-use Filament\Forms\Get;
+use Filament\Schemas\Components\Group;
+use Filament\Schemas\Components\Utilities\Get;
+use Filament\Schemas\Schema;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Str;
+use Livewire\Component;
+use Livewire\Mechanisms\HandleComponents\ComponentContext;
 
 trait HandlesNavigationBuilder
 {
@@ -73,7 +74,7 @@ trait HandlesNavigationBuilder
     {
         return [
             Action::make('item')
-                ->mountUsing(function (ComponentContainer $form) {
+                ->mountUsing(function (Schema $form) {
                     if (! $this->mountedItem) {
                         return;
                     }
@@ -81,7 +82,7 @@ trait HandlesNavigationBuilder
                     $form->fill($this->mountedItemData);
                 })
                 ->view('filament-navigation::hidden-action')
-                ->form([
+                ->schema([
                     TextInput::make('label')
                         ->label(__('filament-navigation::filament-navigation.items-modal.label'))
                         ->required(),
@@ -103,7 +104,7 @@ trait HandlesNavigationBuilder
                             $component
                                 ->getContainer()
                                 ->getComponent(fn (Component $component) => $component instanceof Group)
-                                ->getChildComponentContainer()
+                                ->getChildSchema()
                                 ->fill();
                         })
                         ->reactive(),
@@ -117,8 +118,8 @@ trait HandlesNavigationBuilder
                         }),
                     Group::make()
                         ->statePath('data')
-                        ->visible(fn (Component $component) => $component->evaluate(FilamentNavigation::get()->getExtraFields()) !== [])
-                        ->schema(function (Component $component) {
+                        ->visible(fn (\Filament\Schemas\Components\Component $component) => $component->evaluate(FilamentNavigation::get()->getExtraFields()) !== [])
+                        ->schema(function () {
                             return FilamentNavigation::get()->getExtraFields();
                         }),
                 ])
@@ -149,7 +150,7 @@ trait HandlesNavigationBuilder
 
                     $this->mountedActionData = [];
                 })
-                ->modalButton(__('filament-navigation::filament-navigation.items-modal.btn'))
+                ->modalSubmitActionLabel(__('filament-navigation::filament-navigation.items-modal.btn'))
                 ->label(__('filament-navigation::filament-navigation.items-modal.title')),
         ];
     }
